refactor(types): move MilestoneId + MilestoneRecord to origin-types
Moves the wire types (enum + struct + their as_str/FromStr impls) to the
lightweight origin-types crate so origin-mcp and future plugins can use
them without pulling in origin-core. origin-core re-exports both names via
`pub use origin_types::onboarding::{MilestoneId, MilestoneRecord}` so all
existing callers are unaffected. MilestoneEvaluator stays in origin-core.
